SOme cleanup of DelegationTokenManager
